Chuyển sang làm phần secutiry cho dự án #đã config base hệ thống bảo mật của spring, thêm thư viện jwts cho việc xác thực

+++ src/main/java/com/hustVN/otherShopYet/model/dtos/OrderDetailDTO.java
@@ -14,14 +14,16 @@ public class OrderDetailDTO {
     @JsonProperty("order_id")
     @Min(value = 1, message = "must be > 1")
     private Long orderId;
+
     @JsonProperty("product_id")
     @Min(value = 1, message = "must be > 0")
     private Long productId;
+
     @Min(value = 0, message = "must be greater than or equals 0")
     private Float price;
     @JsonProperty("number_of_products")
     private int numberOfProducts;
 
     @JsonProperty("total_money")
-    private int totalMoney;
+    private Float totalMoney;
 }
